Added setAll() and changed authorId to string instead of UUID on DTO
According to 
http://jsonapi.org/format/

"Client-Generated IDs

A server MAY accept a client-generated ID along with a request to create a resource. An ID MUST be specified with an id key, the value of which MUST be a universally unique identifier. The client SHOULD use a properly generated and formatted UUID as described in RFC 4122 [RFC4122]. "

Thus since the id is already a UUID, the authorId cannot be a UUID, but has to be a String.